[common/datablocks] add blocks assert check macros
